Moved ContractTestSettings into it's own file
